requirements.txt
# onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`

Flask==3.0.3
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.8
langchain-chroma==0.1.4
openai==1.54.4
python-dotenv==1.0.1
chromadb==0.5.0
tiktoken==0.7.0
PyPDF2==3.0.1


